Using ArrayList to store payment items
Datastructures have been reworked to make everything better. We were
using ArrayMap before but things were not in order so that wasn't good
now we are using ArrayList<Pair(String,Boolean)> it keeps things more
organized and in order. Also created my custom EditText so it can store
positions of the answerBoxes to know which value belongs where